refactor(agents): split terminal error boundary
